% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{all_year_summaries}
\alias{all_year_summaries}
\title{all_year_summaries}
\usage{
all_year_summaries(x, sum_col)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}
}
\value{
data frame with 10 columns
}
\description{
Get site summaries
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
summary_info <- all_year_summaries(aquifer_data, sum_col)
}
