% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{chlorideTimeseriesTable}
\alias{chlorideTimeseriesTable}
\title{chlorideTimeSeriesTable}
\usage{
chlorideTimeseriesTable(MainDir, FullSiteName, dat, trendResults_i,
  trendResults_i2, pc)
}
\arguments{
\item{MainDir}{dir}

\item{FullSiteName}{name}

\item{dat}{data frame?}

\item{trendResults_i}{data frame?}

\item{trendResults_i2}{data frame?}

\item{pc}{data frame?}
}
\description{
Not the best name for this funtion - This function creates all the time series tables and plots
not just the chloride time series tables.
}
