% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chloride.R
\name{trend_plot}
\alias{trend_plot}
\title{Chloride plot with trend}
\usage{
trend_plot(
  qw_data,
  plot_title,
  y_label = NA,
  CharacteristicName = c("Chloride"),
  norm_range = c(225, 999),
  include_table = TRUE
)
}
\arguments{
\item{qw_data}{data frame returned from dataRetrieval::readWQPqw,
must include columns ActivityStartDateTime, CharacteristicName, result_va}

\item{plot_title}{character}

\item{CharacteristicName}{character CharacteristicName to plot}

\item{norm_range}{a numerical range to potentially group the data. If NA, no grouping is shown.}

\item{include_table}{logical whether or not to include the trend table in the upper left corner.}
}
\description{
Function creates the cloride over time plot with trends.
}
\examples{

# site <- "263819081585801"
# parameterCd <- c("00095","90095","00940","99220")
# site_data <- dataRetrieval::readWQPqw(site, 
#                                        parameterCd)
# Using package example data:
qw_data <- L2701_example_data$QW
plot_title <- paste(attr(qw_data, "siteInfo")[["station_nm"]], ": Chloride")
trend_plot(qw_data, plot_title)
}
