% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{composite_data}
\alias{composite_data}
\title{Composite hydrograph data}
\usage{
composite_data(x, sum_col, num_years, parameter_cd_gwl = NA)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}

\item{num_years}{integer number of years required}

\item{parameter_cd_gwl}{Parameter code(s) to be filtered to in a column specifically
named "parameter_cd". If the
data doesn't come directly from NWIS services, this can be set to \code{NA},
and this argument will be ignored.}
}
\value{
data frame with year, name, and value
}
\description{
Create composite data
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

comp_data <- composite_data(aquifer_data, sum_col, num_years,
                            parameter_cd_gwl = "72019")

}
