% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groundwater_report.R
\name{create_groundwater_report}
\alias{create_groundwater_report}
\title{Generate HASP report}
\usage{
create_groundwater_report(
  siteID,
  report_name,
  report_folder,
  output_type = "word"
)
}
\arguments{
\item{siteID}{character. USGS site ID, should be a groundwater site.}

\item{report_name}{character. Base name of report.}

\item{report_folder}{character. Report folder, can be full path or partial.}

\item{output_type}{should be either "word", "html", or "pdf".}
}
\description{
Opens a template of the blanks sample report. Running the function
will create a Rmd file. The file can be "knit" as-is, or adjusted before rendering.
}
\examples{


\dontrun{
create_groundwater_report("USGS-253029080295601",
                          report_name = "example_report",
                          report_folder = "reports",
                          output_type = "word")
                          
create_groundwater_report("USGS-424520070562401",
                          report_name = "example_report",
                          report_folder = "reports",
                          output_type = "html")
}

}
