% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_targets_script.R
\name{create_targets_script}
\alias{create_targets_script}
\title{Generate targets script for national aquifer data pull}
\usage{
create_targets_script(
  national_aquifer_cd,
  parameter_cd = "72019",
  statistic_cd = "00003",
  start_date,
  end_date,
  file_name = "_targets.R"
)
}
\arguments{
\item{national_aquifer_cd}{character.}

\item{parameter_cd}{character.}

\item{statistic_cd}{character.}

\item{start_date}{character.}

\item{end_date}{character.}

\item{file_name}{description}
}
\description{
Opens a template of the blanks sample report. Running the function
will create a Rmd file.
}
\examples{


\dontrun{
create_targets_script(national_aquifer_cd = "N100BSNRGB",
                      parameter_cd = "72019",
                      statistic_cd = "00003",
                      start_date = "1988-10-01",
                      end_date = "2021-01-01",
                      file_name = "_targets_test.R")
}

}
