% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_frequency_table}
\alias{daily_frequency_table}
\title{Daily frequency table}
\usage{
daily_frequency_table(
  gw_level_dv,
  parameter_cd = NA,
  date_col = NA,
  value_col = NA,
  approved_col = NA
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data
from \code{readNWISdv}}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd 
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this 
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}
}
\value{
a data frame giving the max, mean, min, and number of available
days of data for each day of the year.
}
\description{
Give the historical max, mean, minimum, and number of available points
for each day of the year
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_frequency_table(gw_level_dv,
                      parameter_cd = "62610")

}
