% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_2yr_plot}
\alias{daily_gwl_2yr_plot}
\title{Plot the last two years of daily data}
\usage{
daily_gwl_2yr_plot(gw_level_dv, parameterCd, statCd, title = "")
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{parameterCd}{the parameter code used}

\item{statCd}{the statistic code used}

\item{title}{the title to use on the plot}
}
\value{
a ggplot object with a ribbon indicating the historical daily range,
green line representing the historical daily median, and approved and provisional
daily data for the last two years
}
\description{
Plot the last two years of daily data
}
\examples{

site <- "263819081585801"
parameterCd <- "62610"
statCd <- "00001"
gwl_data <- dataRetrieval::readNWISdv(site, parameterCd, statCd = statCd)
daily_gwl_2yr_plot(gwl_data, parameterCd, statCd, title = "Groundwater Level")

}
