% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{data_available}
\alias{data_available}
\title{site_summary}
\usage{
data_available(siteID)
}
\arguments{
\item{siteID}{character. USGS site ID for a groundwater site.}
}
\description{
Get station summary information
}
\examples{
siteID <- "263819081585801"
site_data_available <- data_available(siteID)
}
