% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{DataCompletnessAnalyis}
\alias{DataCompletnessAnalyis}
\title{DataCompletnessAnalyis}
\usage{
DataCompletnessAnalyis(clData_five, clData_twenty, pctComplete)
}
\description{
Test to see if there is enough data for the 5
and 20 year trend analyses
}
