% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data.R
\docType{data}
\name{L2701_example_data}
\alias{L2701_example_data}
\title{Example groundwater level data}
\description{
Example data from site 263819081585801 L-2701. Data is a named list of
"Daily" for daily groundwater level at the site and "Discrete" for discrete
groundwater level measurements from the site. Updated April 9, 2021.
}
\examples{
head(L2701_example_data$Daily)
head(L2701_example_data$Discrete)
head(L2701_example_data$QW)
}
