% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{first_day}
\alias{first_day}
\title{Find the first day of the month for a given date}
\usage{
first_day(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
the first day of the month that given dates fall in
}
\description{
Find the first day of the month for a given date
}
\examples{
date <- as.Date("2020-12-28")
first_day(date)

}
