% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{get_range}
\alias{get_range}
\title{Get range}
\usage{
get_range(plot_range, full_range)
}
\arguments{
\item{plot_range}{the time frame to use for the plot. Use "Past year" (default) to see the
last year of data, or "Calendar year" to use the current calendar year, beginning
in January. Or specify two dates representing the start and end of the plot.}

\item{full_range}{the full time range of the data.}
}
\value{
list of plot_start, plot_end, label
}
\description{
Get range
}
\examples{
full_range <- c("2000-01-01", "2025-10-01")
get_range("Past year", full_range)
get_range("Calendar year", full_range)
get_range(c("2024-10-01", "2025-06-01"), full_range)
get_range(c(NA, "2020-01-01"), full_range)
get_range(c("2020-01-01", NA), full_range)
}
