% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_state_data}
\alias{get_state_data}
\title{get_state_data}
\usage{
get_state_data(state, aquiferCd, startDate, endDate, parameter_cd = "72019")
}
\arguments{
\item{state}{character. Can be state abbreviation, long name, or numeric code.}

\item{startDate}{date or string. Beginning date of when to pull data.}

\item{endDate}{date of string  Ending date to pull data.}

\item{parameter_cd}{5-digit character USGS parameter code. Default is "72019".}
}
\description{
Get USGS data based for a single state.
}
\examples{
end_date <- "2021-01-01"
start_date <- "1989-12-31"

\donttest{
st_data <- get_state_data("WI", start_date, end_date)
}
}
