% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{getParmCodeDef}
\alias{getParmCodeDef}
\title{getParmCodeDef}
\usage{
getParmCodeDef(pc)
}
\arguments{
\item{pc}{parameter code}
}
\description{
Function provides the text description for the parameter codes
}
\examples{
pc <- "99220"
getParmCodeDef(pc)
#TODO: replace this function with readNWISpcode!
}
