% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwl_single_sites.R
\name{gwl_plot_periodic}
\alias{gwl_plot_periodic}
\alias{gwl_plot_all}
\title{Single site groundwater level plots and tables}
\usage{
gwl_plot_periodic(gwl, title = "")

gwl_plot_all(dv, gwl, title = "")
}
\arguments{
\item{gwl}{data frame returned from dataRetrieval::readNWISgwl}

\item{title}{character}

\item{dv}{daily value groundwater levels. Must include columns}
}
\description{
Funtion to create the periodic groundwater data plot.
}
\examples{

site <- "263819081585801"
gwl_data <- dataRetrieval::readNWISgwl(site)
title <- attr(gwl_data, "siteInfo")[["station_nm"]]
gwl_plot_periodic(gwl_data, title = title)
site <- "263819081585801"
parameterCd <- "62610"
statCd <- "00001"
dv <- dataRetrieval::readNWISdv(site, parameterCd, statCd = statCd)

gwl_plot_all(dv, gwl_data)
}
