% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data.R
\docType{package}
\name{HASP-package}
\alias{HASP-package}
\title{Analyze groundwater data}
\description{
\code{HASP} includes a set of functions to get groundwater data
as grouped by aquifers and create composite hydrographs.
}
\details{
\tabular{ll}{
Package: \tab HASP\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\keyword{Groundwater}
\keyword{levels}
