% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_level_stats.R
\docType{data}
\name{markerTable}
\alias{markerTable}
\alias{markerTable2}
\title{HASP data included}
\value{
data frame with columns Trend and MarkerDescription

data frame with columns Trend and MarkerDescription
}
\description{
markerTable raw data
}
\examples{
head(markerTable)
head(markerTable2)
}
\keyword{datasets}
