% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{interpretTrend}
\alias{interpretTrend}
\title{interpretTrend}
\usage{
interpretTrend(clTrend, trendLength, trendType)
}
\arguments{
\item{trendType}{}
}
\description{
Something about interpretTrend
}
