% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{kendall_test_5_20_years}
\alias{kendall_test_5_20_years}
\title{Seasonal Kendall Trend Test}
\usage{
kendall_test_5_20_years(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = NA,
  value_col = NA,
  approved_col = NA,
  stat_cd = NA,
  alpha = 0.95,
  seasonal = TRUE,
  enough_5 = 10,
  enough_20 = 6,
  include_current_year = FALSE
)
}
\arguments{
\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd 
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this 
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{stat_cd}{If data in gw_level_dv comes from NWIS, the stat_cd 
can be used to help define the value_col.}

\item{alpha}{the confidence level to use for statistical significance}

\item{seasonal}{logical. Use a seasonal kendall test or not seasonal. Default is \code{TRUE}.}

\item{enough_5}{number per year. Default is 10.}

\item{enough_20}{number per year. Default is 6.}

\item{include_current_year}{a logical indicating whether to include data from
the current calendar year in the test.}

\item{gwl}{data frame that must include a numeric column defined by "value_col",
and a Date or POSIXct column defined by "date_col"}
}
\value{
a data frame of test results from 5 and 20 year Kendall Seasonal Trend test
}
\description{
Test for five and twenty year trends in groundwater level using the 
Seasonal Kendall Trend Test.
}
\details{
At least 10 readings per year for the last 5 years are required for the
5-year test, and at least 6 readings for the last 20 years are required
for the 20-year test. The current calendar year is excluded by default.
}
\examples{

# site <- "260041080093102"
# gw_level_data <- dataRetrieval::readNWISgwl(site)

# Using package example data:
gwl_data <- L2701_example_data$Discrete
kendall_test_5_20_years(NULL,
                        gwl_data,
                        parameter_cd = "62610")
                        
gw_level_dv <- L2701_example_data$Daily
kendall_test_5_20_years(gw_level_dv,
                        gwl_data,
                        parameter_cd = "62610")
}
