% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{kendell_test_5_20_years}
\alias{kendell_test_5_20_years}
\title{Seasonal Kendall Trend Test}
\usage{
kendell_test_5_20_years(
  gwl,
  parameter_cd_gwl = NA,
  date_col = "lev_dt",
  value_col = "sl_lev_va",
  alpha = 0.95,
  seasonal = TRUE,
  enough_5 = 10,
  enough_20 = 6,
  include_current_year = FALSE
)
}
\arguments{
\item{gwl}{data frame that must include a numeric column defined by "value_col",
and a Date or POSIXct column defined by "date_col"}

\item{parameter_cd_gwl}{Parameter code(s) to be filtered to in a column specifically
named "parameter_cd". If the
data doesn't come directly from NWIS services, this can be set to \code{NA},
and this argument will be ignored.s}

\item{date_col}{name of date column.}

\item{value_col}{name of value column.}

\item{alpha}{the confidence level to use for statistical significance}

\item{seasonal}{logical. Use a seasonal kendall test or not seasonal. Default is \code{TRUE}.}

\item{enough_5}{number per year. Default is 10.}

\item{enough_20}{numbr per year. Default is 6.}

\item{include_current_year}{a logical indicating whether to include data from
the current calendar year in the test.}
}
\value{
a data frame of test results from 5 and 20 year Kendall Seasonal Trend test
}
\description{
Test for five and twenty year trends in groundwater level using the 
Seasonal Kendall Trend Test.
}
\details{
At least 10 readings per year for the last 5 years are required for the
5-year test, and at least 6 readings for the last 20 years are required
for the 20-year test. The current calendar year is excluded by default.
}
\examples{

# site <- "260041080093102"
# gw_level_data <- dataRetrieval::readNWISgwl(site)

# Using package example data:
gwl_data <- L2701_example_data$Discrete
kendell_test_5_20_years(gwl_data, parameter_cd_gwl = "62610")

}
