% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{last_day}
\alias{last_day}
\title{Find the last day of the month for a given date}
\usage{
last_day(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
the last day of the month that given dates fall in
}
\description{
Find the last day of the month for a given date
}
