% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_level_stats.R
\name{markerLookup_gw}
\alias{markerLookup_gw}
\title{markerLookup_gw}
\usage{
markerLookup_gw(inputTrend)
}
\arguments{
\item{inputTrend}{character?}
}
\description{
Looks up based on GenericMarkerLookupTable.csv
}
