% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{mid_month}
\alias{mid_month}
\title{Find the middle of the month for a given date}
\usage{
mid_month(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
the middle day of the month the given dates fall in
}
\description{
Find the middle of the month for a given date
}
