% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{monthly_mean}
\alias{monthly_mean}
\title{Monthly mean}
\usage{
monthly_mean(x, date_col = "time", value_col = "value")
}
\arguments{
\item{x}{data.frame}

\item{date_col}{character name of date column}

\item{value_col}{character name of value column}
}
\description{
Take the mean of each month, filter if there's not at least 15 days, 
and create a new date that is the midpoint of each month.
}
\examples{

# site <- "USGS-263819081585801"
p_code_dv <- "62610"
# statCd <- "00001"
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd,
#                                                    skipGeometry = TRUE)
gw_level_dv <- L2701_example_data$Daily

site_statistics <- monthly_frequency_table(gw_level_dv, 
                                           NULL,
                                           parameter_cd = p_code_dv)

gw_monthly <- monthly_mean(gw_level_dv)

}
