% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{normalized_data}
\alias{normalized_data}
\title{Composite normalized hydrograph data}
\usage{
normalized_data(x, sum_col, num_years)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}

\item{num_years}{integer number of years required}
}
\value{
data frame with year, name, and value
}
\description{
Create normalized composite data
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

norm_data <- normalized_data(aquifer_data, sum_col, num_years)
}
