% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{OverallTrend}
\alias{OverallTrend}
\title{OverallTrend}
\usage{
OverallTrend(salinityCol, cltrend_five, cltrend_twenty)
}
\description{
Determine what the overall trend shown on the map is going to be.
If there is a statistically significant trend for one of the periods
and no trend or not enough data to test for the other periods than the 
trend for the available period is shown.
}
\details{
If the trend direction is the same for both periods than the 20 year trend 
symbol is shown.

If the 5 and 20 year trends are in different directions than the double headed 
arrow is shown.
}
