% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{prep_map_data}
\alias{prep_map_data}
\title{prep_map_data}
\usage{
prep_map_data(sites)
}
\arguments{
\item{sites}{aquifer data}
}
\value{
data frame
}
\description{
Get map info
}
\examples{
site_info <- site_info
map_info <- prep_map_data(site_info)
}
\keyword{internal}
