% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data.R
\docType{data}
\name{aquifer_data}
\alias{aquifer_data}
\alias{site_info}
\title{Example aquifer data}
\description{
Example data representing Basin and Range basin-fill aquifers (N100BSNRGB).
}
\examples{
head(aquifer_data)
head(site_info)
}
\keyword{data}
\keyword{quality}
\keyword{water}
