% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_Cl_plot.R
\name{Sc_Cl_plot}
\alias{Sc_Cl_plot}
\title{Sc_CL_plot}
\usage{
Sc_Cl_plot(x, title)
}
\arguments{
\item{x}{data frame}

\item{title}{character}
}
\description{
Function creates the individual chloride versus specfic conductance plots for each site
as well as the chloride versus specific conductance plot for all sites.
}
\examples{

site <- "263819081585801"
parameterCd <- c("00095","90095","00940","99220")
site_data <- dataRetrieval::readNWISqw(site, 
                                       parameterCd)
Sc_Cl_plot(site_data, title = "Hi")
}
