% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_Cl_plot.R
\name{Sc_Cl_table}
\alias{Sc_Cl_table}
\title{Sc_Cl_table}
\usage{
Sc_Cl_table(x)
}
\arguments{
\item{x}{data frame}
}
\description{
Function creates the individual chloride versus specfic conductance tables for each site
as well as the chloride versus specific conductance plot for all sites.
}
\examples{

site <- "263819081585801"
parameterCd <- c("00095","90095","00940","99220")
site_data <- dataRetrieval::readNWISqw(site, 
                                       parameterCd)
sc_cl <- Sc_Cl_table(site_data)
}
