% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{ScClPlotTable}
\alias{ScClPlotTable}
\title{ScCLPlotTable}
\usage{
ScClPlotTable(Plotdata, plotTitle, allData = FALSE)
}
\arguments{
\item{Plotdata}{data frame}

\item{plotTitle}{character}

\item{allData}{logical}
}
\description{
Function creates the individual chloride versus specfic conductance plots for each site
as well as the chloride versus specific conductance plot for all sites.
}
