% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{seasonal_kendall_trend_test}
\alias{seasonal_kendall_trend_test}
\title{Seasonal Kendall Trend Test}
\usage{
seasonal_kendall_trend_test(
  gw_level_data,
  alpha = 0.95,
  include_current_year = FALSE
)
}
\arguments{
\item{gw_level_data}{groundwater level data from the \code{readNWISgwl} from
the dataRetrieval package}

\item{alpha}{the confidence level to use for statistical significance}

\item{include_current_year}{a logical indicating whether to include data from
the current calendar year in the test.}
}
\value{
a data frame of test results from 5 and 20 year Kendall Seasonal Trend test
}
\description{
Test for five and twenty year trends in groundwater level using the 
Seasonal Kendall Trend Test.
}
\details{
At least 10 readings per year for the last 5 years are required for the
5-year test, and at least 6 readings for the last 20 years are required
for the 20-year test. The current calendar year is excluded by default.
}
\examples{

site <- "260041080093102"
gw_level_data <- dataRetrieval::readNWISgwl(site)
seasonal_kendall_trend_test(gw_level_data)

}
