% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{site_summary}
\alias{site_summary}
\title{site_summary}
\usage{
site_summary(siteID, markdown = FALSE)
}
\arguments{
\item{siteID}{character}

\item{markdown}{logical. Use markdown formating or console-friendly.}
}
\description{
Get station summary information
}
\examples{
siteID <- "263819081585801"
site_metadata <- site_summary(siteID)
}
