% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trend_test}
\alias{trend_test}
\title{Trend Test}
\usage{
trend_test(
  gw_level_dv,
  gwl_data,
  n_years = 10,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status"),
  pctComplete = 0.5,
  days_required_per_month = 14,
  POR_trend = TRUE
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{gwl_data}{data frame returned from \code{\link[dataRetrieval]{read_waterdata_field_measurements}}, or
data frame with a date, value, and approval columns. 
Use \code{NULL} for no discrete data.}

\item{n_years}{integer. This is the number of years to calculate the trend on.
Default is 10. This can be a vector of years.}

\item{parameter_cd}{Can be used to filter data if the data frame has a 
"parameter_code" column. The default is \code{NA},
which will not do any filtering. If the gwl_data and gw_level_dv need different
parameter code filtering, use a vector of 2 parameter codes. The first one will filter the
gw_level_dv data frame, the second will filter the gwl_data data frame.}

\item{date_col}{the name of the time columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("time", "time")}.}

\item{value_col}{the name of the value columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("value", "value")}.}

\item{approved_col}{the name of the column to get provisional/approved status. 
The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("approval_status", "approval_status")}.
It is expected that these columns will have only "Approved" or "Provisional".}

\item{pctComplete}{number percentage complete. This is a fraction that represents
the amount of data that must be included overall in order to calculate a trend.
The default is 0.5, which means if gaps in the data span more than 50% of the
total record, a trend will not be calculated.}

\item{days_required_per_month}{integer. Number of days required per month
to include in the trend test. Default is 14.}

\item{POR_trend}{a logical indicating whether to include a trend test
for the full period of record. Default is \code{TRUE}.}
}
\value{
a data frame of test results from 5 and 20 year Kendall Seasonal Trend test
}
\description{
Test for period of record and user-specified ranges. The default
trends are calculated for 10 year and the full period of record.
}
\details{
For data that is at least on a daily interval, the \link[rkt]{rkt} function
is used. For periodic data, the \link[EnvStats]{kendallTrendTest} is used.
}
\examples{

site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"

# Using package example data:
# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site)
gwl_data <- L2701_example_data$Discrete
                        
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd)
                                                    
gw_level_dv <- L2701_example_data$Daily

trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610")
           
trend_test(gw_level_dv,
           gwl_data,
           POR_trend = FALSE,
           parameter_cd = "62610")
           
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610",
           n_years = 5)
           
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610",
           n_years = c(5, 10, 20))

# Only periodic data:
trend_test(NULL,
           gwl_data,
           parameter_cd = "62610")
           
}
