% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{trendAnalysis}
\alias{trendAnalysis}
\title{trendAnalysis}
\usage{
trendAnalysis(siteID, stationName, clData_raw, pctComplete)
}
\arguments{
\item{siteID}{character}

\item{stationName}{character}

\item{clData_raw}{data frame?}

\item{pctComplete}{number?}
}
\description{
This is the main chloride trend analysis script, but it is very modular in design
it calls a series of other funtions that conduct parts of the analysis,
}
