% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{trendTest}
\alias{trendTest}
\title{trendTest}
\usage{
trendTest(trendType, salinityCol, clData = NULL, pctComplete = NULL,
  cltrend_five = NULL, cltrend_twenty = NULL, MainDir, DataCmpltAnaly)
}
\arguments{
\item{DataCmpltAnaly}{}
}
\description{
This trend test funciton calls the kendal trend test function developed by the EPA. The season kendal trend
trend could also have been used but much of the data is annual and there is limited seasonality to this data. 
In most instances the overwhelming feature is inland movement of the front rather than seasonality.
}
