% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{water_year}
\alias{water_year}
\title{Convert to water year}
\usage{
water_year(x)
}
\arguments{
\item{x}{character vector}
}
\description{
This function is a little more robust than \code{\link[dataRetrieval]{calcWaterYear}}
}
\examples{
x <- c("2010-01-01", "1994-02", "1980", "2009-11-01")
water_year(x)
}
