% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_plot}
\alias{weekly_frequency_plot}
\title{Plot weekly frequency analysis}
\usage{
weekly_frequency_plot(
  gw_level_dv,
  date_col,
  value_col,
  approved_col,
  plot_range = c("Past year", "Calendar year"),
  plot_title = "",
  y_axis_label = ""
)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{date_col}{name of date column.}

\item{value_col}{name of value column.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{plot_range}{the time frame to use for the plot. Either "Past year" to use the
last year of data, or "Calendar year" to use the current calendar year, beginning
in January.}

\item{plot_title}{the title to use on the plot}

\item{y_axis_label}{the label used for the y-axis of the plot.}
}
\value{
a ggplot object with rectangles representing the historical weekly percentiles,
and points representing the historical median and daily values
}
\description{
The weekly frequency analysis is based on daily data
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
weekly_frequency_plot(gw_level_dv, 
                      date_col = "Date",
                      value_col = "X_62610_00001",
                      approved_col = "X_62610_00001_cd")

}
