% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_table}
\alias{weekly_frequency_table}
\title{Create a table of weekly frequency analysis}
\usage{
weekly_frequency_table(gwl_data, date_col, value_col, approved_col)
}
\arguments{
\item{gwl_data}{daily groundwater level data
from readNWISdv}

\item{date_col}{name of date column.}

\item{value_col}{name of value column.}

\item{approved_col}{name of column to get provisional/approved status.}
}
\value{
a data frame of weekly frequency analysis
}
\description{
The weekly frequency analysis is based on daily values
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gwl_data <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gwl_data <- L2701_example_data$Daily
weekly_frequency <- weekly_frequency_table(gwl_data,
                                           date_col = "Date",
                                           value_col = "X_62610_00001",
                                           approved_col = "X_62610_00001_cd")
}
