% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chloride.R
\name{trend_plot}
\alias{trend_plot}
\title{Chloride plot with trend}
\usage{
trend_plot(
  qw_data,
  plot_title,
  pcode = c("00940", "99220"),
  norm_range = c(225, 999)
)
}
\arguments{
\item{qw_data}{data frame returned from dataRetrieval::readNWISqw,
must include columns sample_dt, parm_cd, result_va}

\item{plot_title}{character}

\item{pcode}{character pcode to plot}

\item{norm_range}{a numerical range to potentially group the data. If NA, no grouping is shown.}
}
\description{
Function creates the cloride over time plot with trends.
}
\examples{

# site <- "263819081585801"
# parameterCd <- c("00095","90095","00940","99220")
# site_data <- dataRetrieval::readNWISqw(site, 
#                                        parameterCd)
# Using package example data:
qw_data <- L2701_example_data$QW
title <- paste(attr(qw_data, "siteInfo")[["station_nm"]], ": Chloride")
trend_plot(qw_data, plot_title = title)
}
