% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{composite_data}
\alias{composite_data}
\title{Composite hydrograph data}
\usage{
composite_data(x, num_years)
}
\arguments{
\item{x}{aquifer data}

\item{num_years}{integer number of years required}
}
\value{
data frame with year, name, and value
}
\description{
Create composite data
}
\examples{
aquifer_data <- aquifer_data
num_years <- 30

comp_data <- composite_data(aquifer_data, num_years)

}
