% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_frequency_table}
\alias{daily_frequency_table}
\title{Daily frequency table}
\usage{
daily_frequency_table(gw_level_dv, date_col, value_col, approved_col)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data
from \code{readNWISdv}}

\item{date_col}{the heading of the date column.}

\item{value_col}{name of value column.}

\item{approved_col}{name of column to get provisional/approved status.}
}
\value{
a data frame giving the max, mean, min, and number of available
days of data for each day of the year.
}
\description{
Give the historical max, mean, minimum, and number of available points
for each day of the year
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_frequency_table(gw_level_dv,
                      date_col = "Date",
                      value_col = "X_62610_00001",
                      approved_col = "X_62610_00001_cd")

}
