% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_frequency_table}
\alias{daily_frequency_table}
\title{Daily frequency table}
\usage{
daily_frequency_table(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "lev_dt"),
  value_col = "value",
  approved_col = c("approval_status", "lev_age_cd"),
  stat_cd = NA
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data
from \code{read_waterdata_daily}}

\item{gwl_data}{data frame returned from dataRetrieval::readNWISgwl, or
data frame with mandatory columns lev_dt (representing date), lev_age_cd (representing
approval code), and a column representing the measured value (either lev_va,
sl_lev_va, or value).}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{stat_cd}{If data in gw_level_dv comes from NWIS, the stat_cd
can be used to help define the value_col.}
}
\value{
a data frame giving the max, mean, min, and number of available
days of data for each day of the year.
}
\description{
Calculates the historical max, mean, minimum, and number of available points
for each day of the year
}
\examples{

# site <- "USGS-263819081585801"
p_code_dv <- "62610"
# statCd <- "00001"
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd,
#                                                    skipGeometry = TRUE)
gw_level_dv <- L2701_example_data$Daily
daily_frequency_table(gw_level_dv,
                      NULL,
                      parameter_cd = "62610")

gwl_data <- L2701_example_data$Discrete
daily_frequency_table(gw_level_dv,
                      gwl_data,
                      parameter_cd = "62610")
}
