% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_frequency_table}
\alias{daily_frequency_table}
\title{Daily frequency table}
\usage{
daily_frequency_table(gw_level_dv, p_code_dv, statCd, date_col = "Date")
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{p_code_dv}{the parameter code used}

\item{statCd}{the statistic code used}

\item{date_col}{the heading of the date column, default is "Date"}
}
\value{
a data frame giving the max, mean, min, and number of available
days of data for each day of the year.
}
\description{
Give the historical max, mean, minimum, and number of available points
for each day of the year
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_frequency_table(gw_level_dv, p_code_dv, statCd)

}
