% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_2yr_plot}
\alias{daily_gwl_2yr_plot}
\title{Plot the last two years of daily data}
\usage{
daily_gwl_2yr_plot(
  gw_level_dv,
  parameter_cd = NA,
  date_col = NA,
  value_col = NA,
  approved_col = NA,
  historical_stat = c("mean", "median"),
  month_breaks = FALSE,
  plot_title = "",
  subtitle = "U.S. Geological Survey",
  y_axis_label = "",
  flip_y = FALSE
)
}
\arguments{
\item{gw_level_dv}{data frame daily groundwater level data
from \code{readNWISdv}}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd 
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this 
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{historical_stat}{the summary statstic to use for middle line of the plot. Either
"mean" or "median."}

\item{month_breaks}{a logical indicating whether to use monthly breaks for the plot}

\item{plot_title}{the title to use on the plot}

\item{subtitle}{character. Sub-title for plot, default is "U.S. Geological Survey".}

\item{y_axis_label}{the label to use for the y axis}

\item{flip_y}{logical. If \code{TRUE}, flips the y axis so that the smallest number is on top.
Default is \code{FALSE}.}
}
\value{
a ggplot object with a ribbon indicating the historical daily range,
the historical daily mean or median, and approved and provisional
daily data for the last two years
}
\description{
Plot the last two years of daily data
}
\examples{

site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_gwl_2yr_plot(gw_level_dv,
                   parameter_cd = "62610",
                   plot_title = "Groundwater Level", 
                   month_breaks = TRUE,
                   historical_stat = "median")

daily_gwl_2yr_plot(gw_level_dv,
                   date_col = "Date",
                   value_col = "X_62610_00001",
                   approved_col = "X_62610_00001_cd", 
                   plot_title = "Groundwater Level", 
                   month_breaks = TRUE,
                   historical_stat = "median",
                   flip_y = FALSE)
}
