% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_2yr_plot}
\alias{daily_gwl_2yr_plot}
\title{Plot the last two years of daily data}
\usage{
daily_gwl_2yr_plot(
  gw_level_dv,
  p_code_dv,
  statCd,
  date_col = "Date",
  plot_title = "",
  historical_stat = c("mean", "median"),
  month_breaks = FALSE
)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{p_code_dv}{the parameter code used}

\item{statCd}{the statistic code used}

\item{date_col}{name of date column. Default is "Date"}

\item{plot_title}{the title to use on the plot}

\item{historical_stat}{the summary statstic to use for middle line of the plot. Either
"mean" or "median."}

\item{month_breaks}{a logical indicating whether to use monthly breaks for the plot}
}
\value{
a ggplot object with a ribbon indicating the historical daily range,
the historical daily mean or median, and approved and provisional
daily data for the last two years
}
\description{
Plot the last two years of daily data
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_gwl_2yr_plot(gw_level_dv, p_code_dv, statCd, 
                   plot_title = "Groundwater Level", 
                   month_breaks = TRUE,
                   historical_stat = "median")

}
