% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_2yr_plot}
\alias{daily_gwl_2yr_plot}
\title{Plot the last two years of daily data}
\usage{
daily_gwl_2yr_plot(
  gw_level_dv,
  date_col,
  value_col,
  approved_col,
  historical_stat = c("mean", "median"),
  month_breaks = FALSE,
  plot_title = "",
  y_axis_label = "",
  flip_y = TRUE
)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{date_col}{name of date column.}

\item{value_col}{name of value column.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{historical_stat}{the summary statstic to use for middle line of the plot. Either
"mean" or "median."}

\item{month_breaks}{a logical indicating whether to use monthly breaks for the plot}

\item{plot_title}{the title to use on the plot}

\item{y_axis_label}{the label to use for the y axis}

\item{flip_y}{logical. If \code{TRUE}, flips the y axis so that the smallest number is on top.
Default is \code{TRUE}.}
}
\value{
a ggplot object with a ribbon indicating the historical daily range,
the historical daily mean or median, and approved and provisional
daily data for the last two years
}
\description{
Plot the last two years of daily data
}
\examples{

site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_gwl_2yr_plot(gw_level_dv,
                   date_col = "Date",
                   value_col = "X_62610_00001",
                   approved_col = "X_62610_00001_cd", 
                   plot_title = "Groundwater Level", 
                   month_breaks = TRUE,
                   historical_stat = "median")

daily_gwl_2yr_plot(gw_level_dv,
                   date_col = "Date",
                   value_col = "X_62610_00001",
                   approved_col = "X_62610_00001_cd", 
                   plot_title = "Groundwater Level", 
                   month_breaks = TRUE,
                   historical_stat = "median",
                   flip_y = FALSE)
}
