% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_plot}
\alias{daily_gwl_plot}
\title{Plot recent data}
\usage{
daily_gwl_plot(
  gw_level_dv,
  parameter_cd = NA,
  date_col = NA,
  value_col = NA,
  approved_col = NA,
  stat_cd = NA,
  start_date = NA,
  end_date = NA,
  historical_stat = "mean",
  month_breaks = FALSE,
  plot_title = "",
  subtitle = "U.S. Geological Survey",
  y_axis_label = "",
  flip = FALSE
)
}
\arguments{
\item{gw_level_dv}{data frame daily groundwater level data
from \code{readNWISdv}}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd 
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this 
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{stat_cd}{If data in gw_level_dv comes from NWIS, the stat_cd 
can be used to help define the value_col.}

\item{start_date}{Date to start plot. If \code{NA} (which is the default),
the plot will start 2 years before the most recent value.}

\item{historical_stat}{the summary statistic to use for middle line of the plot. Either
"mean" or "median."}

\item{month_breaks}{a logical indicating whether to use monthly breaks for the plot}

\item{plot_title}{the title to use on the plot}

\item{subtitle}{character. Sub-title for plot, default is "U.S. Geological Survey".}

\item{y_axis_label}{the label to use for the y axis}

\item{flip}{logical. If \code{TRUE}, flips the y axis so that the smallest number is on top.
Default is \code{FALSE}.}

\item{end_data}{Date to end plot. If \code{NA} (which is the default), 
the plot will end with the latest measurement.}
}
\value{
a ggplot object with a ribbon indicating the historical daily range,
the historical daily mean or median, and approved and provisional
daily data for the last two years
}
\description{
Plot recent data
}
\examples{

site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_gwl_plot(gw_level_dv,
               parameter_cd = "62610",
               plot_title = "Groundwater Level", 
               month_breaks = TRUE,
               historical_stat = "median")
               
daily_gwl_plot(gw_level_dv,
               parameter_cd = "62610",
               plot_title = "Groundwater Level", 
               month_breaks = TRUE,
               start_date = "2020-10-01",
               historical_stat = "median")
               
daily_gwl_plot(gw_level_dv,
               parameter_cd = "62610",
               plot_title = "Groundwater Level", 
               month_breaks = TRUE,
               start_date = "2018-10-01",
               end_date = "2020-10-01",
               historical_stat = "median")

}
