% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_summary}
\alias{daily_gwl_summary}
\title{Summary table of daily data}
\usage{
daily_gwl_summary(gw_level_dv, parameterCd, statCd)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{parameterCd}{the parameter code used}

\item{statCd}{the statistic code used}
}
\value{
a summary table giving the period of record, completeness
and percentile values
}
\description{
Summary table of daily data
}
\examples{

# site <- "263819081585801"
parameterCd <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, parameterCd, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
daily_gwl_summary(gw_level_dv, parameterCd, statCd)

}
