% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{daily_gwl_summary}
\alias{daily_gwl_summary}
\title{Summary table of daily data}
\usage{
daily_gwl_summary(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status")
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{gwl_data}{data frame returned from \code{\link[dataRetrieval]{read_waterdata_field_measurements}}, or
data frame with a date, value, and approval columns. 
Use \code{NULL} for no discrete data.}

\item{parameter_cd}{Can be used to filter data if the data frame has a 
"parameter_code" column. The default is \code{NA},
which will not do any filtering.}

\item{date_col}{the name of the time columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("time", "time")}.}

\item{value_col}{the name of the value columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("value", "value")}.}

\item{approved_col}{the name of the column to get provisional/approved status. 
The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("approval_status", "approval_status")}.
It is expected that these columns will have only "Approved" or "Provisional".}
}
\value{
a summary table giving the period of record, completeness
and percentile values
}
\description{
Summary table of daily data
}
\examples{

site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd,
#                                                    skipGeometry = TRUE)
#
                                                    
gw_level_dv <- L2701_example_data$Daily

daily_gwl_summary(gw_level_dv,
                  gwl_data = NULL,
                  parameter_cd = p_code_dv)

# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site, 
#                                  skipGeometry = TRUE)
gwl_data <- L2701_example_data$Discrete

daily_gwl_summary(gw_level_dv,
                  gwl_data = gwl_data,
                  parameter_cd = p_code_dv)
}
