% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{filter_sites}
\alias{filter_sites}
\title{filter_sites}
\usage{
filter_sites(x, sum_col, num_years)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}

\item{num_years}{integer number of years required}
}
\value{
data frame filter of x
}
\description{
Filter down to sites with num_years of data
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

aq_data <- filter_sites(aquifer_data, sum_col, num_years)
}
