% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{filter_sites}
\alias{filter_sites}
\title{filter_sites}
\usage{
filter_sites(
  x,
  parameter_cd = "72019",
  num_years = NA,
  start_year = NA,
  end_year = NA
)
}
\arguments{
\item{x}{aquifer data}

\item{parameter_cd}{character, 5-digit parameter code, default is "72019".}

\item{num_years}{integer number of years required. This can be
\code{NA}, in which case the filter will use the full range of the data.}

\item{start_year}{integer the first year to filter from. If \code{NA},
the filter will use the minimum from the data.}

\item{end_year}{integer the last year to filter from. If \code{NA},
the filter will use the last year.}
}
\value{
data frame filter of x
}
\description{
Filter down to sites with num_years of data
}
\examples{
aquifer_data <- aquifer_data
num_years <- 30

aq_data <- filter_sites(aquifer_data,
                        parameter_cd = "72019",
                        num_years = num_years)
}
