% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_aquifer_data}
\alias{get_aquifer_data}
\title{get_aquifer_data}
\usage{
get_aquifer_data(aquiferCd, startDate, endDate)
}
\arguments{
\item{aquiferCd}{character}

\item{startDate}{date or string}

\item{endDate}{date of string}
}
\description{
Get USGS data based on aquiferCd
}
\examples{
end_date <- "2019-12-31"
start_date <- "1989-12-31"

aquiferCd <- "S100CSLLWD"
#aq_data <- get_aquifer_data(aquiferCd, start_date, end_date)
}
