% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwl_single_sites.R
\name{gwl_plot_periodic}
\alias{gwl_plot_periodic}
\alias{gwl_plot_all}
\title{Single site groundwater level plots and tables}
\usage{
gwl_plot_periodic(
  gwl,
  plot_title = "",
  date_col = "lev_dt",
  value_col = "sl_lev_va",
  approved_col = "lev_age_cd"
)

gwl_plot_all(dv, gwl, plot_title = "", add_trend = FALSE)
}
\arguments{
\item{gwl}{data frame returned from dataRetrieval::readNWISgwl}

\item{plot_title}{character}

\item{date_col}{name of date column. Default is "lev_dt".}

\item{value_col}{name of value column. Default is "sl_lev_va".}

\item{approved_col}{name of column to get provisional/approved status.
Default is "lev_age_cd".}

\item{dv}{daily value groundwater levels. Must include columns}

\item{add_trend}{logical. Uses \code{kendell_test_5_20_years}.}
}
\description{
Funtion to create the periodic groundwater data plot.
}
\examples{

# site <- "263819081585801"
# gwl_data <- dataRetrieval::readNWISgwl(site)

# Using package example data:
gwl_data <- L2701_example_data$Discrete
title <- attr(gwl_data, "siteInfo")[["station_nm"]]
gwl_plot_periodic(gwl_data, plot_title = title)
# site <- "263819081585801"
# parameterCd <- "62610"
# statCd <- "00001"
# dv <- dataRetrieval::readNWISdv(site, parameterCd, statCd = statCd)
# Using package example data:
dv <- L2701_example_data$Daily
gwl_plot_all(dv, gwl_data)
}
