% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwl_single_sites.R
\name{gwl_plot_periodic}
\alias{gwl_plot_periodic}
\alias{gwl_summary_periodic}
\title{Single site groundwater level plots and tables}
\usage{
gwl_plot_periodic(x, title = "")

gwl_summary_periodic(x)
}
\arguments{
\item{x}{data frame returned from dataRetrieval::readNWISgwl}

\item{title}{character}
}
\description{
Funtion to create the periodic groundwater data plot.
}
\examples{

site <- "263819081585801"
gwl_data <- dataRetrieval::readNWISgwl(site)
title <- attr(gwl_data, "siteInfo")[["station_nm"]]
gwl_plot_periodic(gwl_data, title = title)
site <- "263819081585801"
gw_gwl_data <- dataRetrieval::readNWISgwl(site)
}
