% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwl_single_sites.R
\name{gwl_plot_field}
\alias{gwl_plot_field}
\alias{gwl_plot_all}
\title{Single site groundwater level plots and tables}
\usage{
gwl_plot_field(
  gwl_data,
  plot_title = "",
  parameter_cd = NA,
  date_col = "time",
  value_col = "value",
  approved_col = "approval_status",
  flip = TRUE,
  y_label = "",
  subtitle = "U.S. Geological Survey"
)

gwl_plot_all(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status"),
  y_label = "",
  subtitle = "U.S. Geological Survey",
  plot_title = "",
  add_trend = FALSE,
  n_years = 10,
  POR_trend = TRUE,
  flip = FALSE,
  days_required_per_month = 14
)
}
\arguments{
\item{gwl_data}{data frame returned from \code{\link[dataRetrieval]{read_waterdata_field_measurements}}, or
data frame with a date, value, and approval columns. 
Use \code{NULL} for no discrete data.}

\item{plot_title}{character, title for plot.}

\item{parameter_cd}{Can be used to filter data if the data frame has a 
"parameter_code" column. The default is \code{NA},
which will not do any filtering.}

\item{date_col}{the name of the time columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("time", "time")}.}

\item{value_col}{the name of the value columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("value", "value")}.}

\item{approved_col}{the name of the column to get provisional/approved status. 
The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("approval_status", "approval_status")}.
It is expected that these columns will have only "Approved" or "Provisional".}

\item{flip}{logical. If \code{TRUE}, flips labels so that the lower numbers
are in the higher percentages. Default is \code{TRUE}.}

\item{y_label}{character for y-axis label. Consider using 
\code{\link[dataRetrieval]{read_waterdata_parameter_codes}} for USGS parameter_name.}

\item{subtitle}{character. Sub-title for plot, default is "U.S. Geological Survey".}

\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{add_trend}{logical. Uses \code{trend_test}.}

\item{n_years}{integer. This is the number of years to calculate the trend on.
Default is 10. This can be a vector of years.}

\item{POR_trend}{a logical indicating whether to include a trend test
for the full period of record. Default is \code{TRUE}.}

\item{days_required_per_month}{integer. Number of days required per month. 
Default is 14. Only used if add_trend is \code{TRUE} using daily data.}
}
\description{
Function to create the field groundwater level data plot.
}
\examples{

site <- "USGS-263819081585801"
# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site, 
#                                  skipGeometry = TRUE)
gwl_data <- L2701_example_data$Discrete

pcodes <- dataRetrieval::read_waterdata_parameter_codes(
            parameter_code = unique(gwl_data$parameter_code)
          )

gwl_plot_field(gwl_data, 
               plot_title = "L2701_example", 
               parameter_cd = "62610",
               flip = FALSE,
               y_label = pcodes$parameter_name[pcodes$parameter_code == "62610"])
               
gwl_plot_field(gwl_data, 
               plot_title = "L2701_example",
               y_label = pcodes$parameter_name[pcodes$parameter_code == "62611"], 
               parameter_cd = "62611",
               flip = FALSE)
               
gwl_plot_field(gwl_data,  
               plot_title = "L2701_example",
               y_label = pcodes$parameter_name[pcodes$parameter_code == "72019"], 
               parameter_cd = "72019",
               flip = TRUE)
               
site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd,
#                                                    skipGeometry = TRUE)
#
                                                    
gw_level_dv <- L2701_example_data$Daily

site_info <- dataRetrieval::read_waterdata_monitoring_location(monitoring_location_id = site)

plot_title <- site_info$monitoring_location_name
pcodes <- dataRetrieval::read_waterdata_parameter_codes(parameter_code = p_code_dv)

gwl_plot_all(gw_level_dv, 
             NULL, 
             plot_title = plot_title,
             y_label = pcodes$parameter_name,
             flip = TRUE) 

# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site, 
#                                  skipGeometry = TRUE)
gwl_data <- L2701_example_data$Discrete

gwl_plot_all(gw_level_dv, 
             gwl_data, 
             parameter_cd = "62610",
             plot_title = plot_title,
             y_label = pcodes$parameter_name,
             add_trend = FALSE,
             flip = FALSE)

gwl_plot_all(gw_level_dv,
             gwl_data,
             parameter_cd = "62610",
             n_years = c(5, 20),
             POR_trend = TRUE,
             y_label = pcodes$parameter_name,
             plot_title = plot_title,
             add_trend = TRUE)
             
gwl_plot_all(NULL, 
             gwl_data, 
             parameter_cd = "62610",
             plot_title = plot_title, 
             y_label = pcodes$parameter_name)

gwl_plot_all(NULL, 
             gwl_data, 
             parameter_cd = "62610",
             plot_title = plot_title,
             y_label = pcodes$parameter_name,
             add_trend = TRUE)

}
