% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data.R
\docType{data}
\name{nat_aqfr_state}
\alias{nat_aqfr_state}
\alias{summary_aquifers}
\alias{local_aqfr}
\alias{markerTable}
\alias{markerTable2}
\title{Aquifer data included}
\value{
data frame with columns Country, State, nat_aqfr_cd, long_name, and other_cd

data frame with columns long_name, nat_aqfr_cd, state_indexes, and states

data frame with aqfr_cd, Aqfr_Name_prpr

data frame with columns Trend and MarkerDescription

data frame with columns trendType, trend, markerDescription, r_lwd
}
\description{
National aquifer data

Local aquifer data codes

markerTable raw data
}
\examples{
head(nat_aqfr_state)
head(summary_aquifers)
head(local_aqfr)
}
\keyword{datasets}
