% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data.R
\docType{data}
\name{nat_aqfr_state}
\alias{nat_aqfr_state}
\alias{summary_aquifers}
\title{Aquifer data included}
\value{
data frame with columns Country, State, nat_aqfr_cd, long_name, and other_cd

data frame with columns long_name, nat_aqfr_cd, state_indexes, and states
}
\description{
National aquifer data
}
\examples{
head(nat_aqfr_state)
head(summary_aquifers)
}
\keyword{datasets}
