% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{last_day}
\alias{last_day}
\title{Find the last day of the month for a given date}
\usage{
last_day(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
the last day of the month that given dates fall in
}
\description{
Find the last day of the month for a given date
}
\examples{
date <- as.Date("2020-12-28")
last_day(date)
last_day("2020-02-15")
last_day("2019-02-15")
last_day(c("2020-12-28", "2020-02-15", "2019-02-15"))
}
