% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_hydro_data}
\alias{map_hydro_data}
\title{Map data used in composite hydrographs}
\usage{
map_hydro_data(x, sum_col, num_years, parameter_cd_gwl)
}
\arguments{
\item{x}{aquifer data frame. Requires at least 3 columns. Two are required "site_no", "year",
and the 3rd is defined by the sum_col argument.}

\item{sum_col}{column name to do the analysis on. In data coming from 
\code{dataRetrieval}, this is often either "sl_lev_va" or "lev_va".}

\item{num_years}{integer number of years required}
}
\value{
leaflet object
}
\description{
Map data used in composite hydrographs
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

map_data <- map_hydro_data(aquifer_data, sum_col, num_years, "72019")
map_data
}
