% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{mid_month}
\alias{mid_month}
\title{Find the middle of the month for a given date}
\usage{
mid_month(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
the middle day of the month the given dates fall in
}
\description{
Find the middle of the month for a given date
}
\examples{
date <- as.Date("2020-12-28")
mid_month(date)
mid_month(c("2019-02-15", "2020-03-08", "2010-06-01"))
}
