% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{monthly_frequency_table}
\alias{monthly_frequency_table}
\title{Create a table of monthly frequency analysis}
\usage{
monthly_frequency_table(gw_level_dv, date_col, value_col, approved_col)
}
\arguments{
\item{gw_level_dv}{groundwater level data from \code{readNWISgwl}}

\item{date_col}{name of date column.}

\item{value_col}{name of value column.}

\item{approved_col}{name of column to get provisional/approved status.}
}
\value{
a data frame of monthly groundwater level statistics including the
5th, 10th, 25th, 75th, 90th, and 95th percentiles; the number of
years of data; and the lowest monthly median and the highest monthly
median.
}
\description{
Create a table of monthly frequency analysis
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
monthly_frequency <- monthly_frequency_table(gw_level_dv,
                                             date_col = "Date",
                                             value_col = "X_62610_00001",
                                             approved_col = "X_62610_00001_cd")

}
