% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{monthly_frequency_table}
\alias{monthly_frequency_table}
\title{Create a table of monthly frequency analysis}
\usage{
monthly_frequency_table(gw_level_data)
}
\arguments{
\item{gw_level_data}{groundwater level data from \code{readNWISgwl}}
}
\value{
a data frame of monthly groundwater level statistics including the
5th, 10th, 25th, 75th, 90th, and 95th percentiles; the number of
years of data; and the lowest monthly median and the highest monthly
median.
}
\description{
Create a table of monthly frequency analysis
}
\examples{

site <- "263819081585801"
gw_level_data <- dataRetrieval::readNWISgwl(site)
monthly_frequency <- monthly_frequency_table(gw_level_data)
}
