% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{monthly_mean}
\alias{monthly_mean}
\title{Monthly mean}
\usage{
monthly_mean(
  x,
  date_col = "Date",
  value_col = "X_62610_00001",
  days_required_per_month = 14
)
}
\arguments{
\item{x}{data.frame}

\item{date_col}{character name of date column}

\item{value_col}{character name of value column}

\item{days_required_per_month}{integer. Number of days required per month. 
Default is 14.}
}
\description{
Take the mean of each month, filter if there's not at least 15 days, 
and create a new date that is the midpoint of each month.
}
\examples{

# site <- "263819081585801"
parameterCd <- "62610"
# statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, parameterCd, 
#                                           statCd = statCd)
# Using package example data:
gw_level_dv <- L2701_example_data$Daily

site_statistics <- monthly_frequency_table(gw_level_dv, 
                                           NULL,
                                           parameter_cd = parameterCd)

gw_monthly <- monthly_mean(gw_level_dv)

}
