% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_composite_data}
\alias{plot_composite_data}
\title{Plot composite hydrograph data}
\usage{
plot_composite_data(x, sum_col, num_years)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}

\item{num_years}{integer number of years required}
}
\value{
ggplot2 object
}
\description{
Create composite hydrograph plot
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

comp_data <- plot_composite_data(aquifer_data, sum_col, num_years)
comp_data
}
