% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_composite_data}
\alias{plot_composite_data}
\title{Plot composite hydrograph data}
\usage{
plot_composite_data(x, sum_col, num_years, plot_title = "")
}
\arguments{
\item{x}{aquifer data frame. Requires at least 3 columns. Two are required "site_no", "year",
and the 3rd is defined by the sum_col argument.}

\item{sum_col}{column name to do the analysis on. In data coming from 
\code{dataRetrieval}, this is often either "sl_lev_va" or "lev_va".}

\item{num_years}{integer number of years required}

\item{plot_title}{character title included on plot.}
}
\value{
ggplot2 object
}
\description{
Create composite hydrograph plot
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
num_years <- 30

comp_data <- plot_composite_data(aquifer_data, sum_col, num_years)
comp_data
#' # Do it on a water year:
aquifer_data$cal_year <- aquifer_data$year
aquifer_data$year <- aquifer_data$water_year
plot_composite_data(aquifer_data, sum_col, num_years)
}
