% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_normalized_data}
\alias{plot_normalized_data}
\title{Plot normalized composite hydrograph data}
\usage{
plot_normalized_data(x, num_years = NA, plot_title = "")
}
\arguments{
\item{x}{aquifer data frame. Requires at least 3 columns. Two are required "site_no", "year",
and "value".}

\item{num_years}{integer number of years required to the analysis. If \code{NA}, the 
analysis will default to the range of the data in x.}

\item{plot_title}{character title of plot.}
}
\value{
ggplot2 object
}
\description{
Create composite hydrograph plot
}
\examples{
aquifer_data <- aquifer_data
num_years <- 30

norm_data <- plot_normalized_data(aquifer_data, num_years)
norm_data

aquifer_data$cal_year <- aquifer_data$year
aquifer_data$year <- aquifer_data$water_year
plot_normalized_data(aquifer_data, num_years = num_years)
}
