% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{prep_map_data}
\alias{prep_map_data}
\title{prep_map_data}
\usage{
prep_map_data(x)
}
\arguments{
\item{x}{aquifer data}
}
\value{
data frame
}
\description{
Get map info
}
\examples{
aquifer_data <- aquifer_data
map_info <- prep_map_data(aquifer_data)
}
