% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{prep_map_data}
\alias{prep_map_data}
\title{prep_map_data}
\usage{
prep_map_data(x, sum_col)
}
\arguments{
\item{x}{aquifer data}

\item{sum_col}{column name}
}
\value{
data frame
}
\description{
Get map info
}
\examples{
aquifer_data <- aquifer_data
sum_col <- "lev_va"
map_info <- prep_map_data(aquifer_data, sum_col)
}
