% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_Cl_plot.R
\name{Sc_Cl_plot}
\alias{Sc_Cl_plot}
\alias{Sc_Cl_table}
\title{Specific conductance and chloride}
\usage{
Sc_Cl_plot(x, title)

Sc_Cl_table(x)
}
\arguments{
\item{x}{data frame returned from dataRetrieval::readNWISqw,
must include columns sample_dt, parm_cd, result_va}

\item{title}{character}
}
\description{
Function creates the individual chloride versus specfic conductance plots and tables for each site
as well as the chloride versus specific conductance plot for all sites.
}
\examples{

# site <- "263819081585801"
# parameterCd <- c("00095","90095","00940","99220")
# site_data <- dataRetrieval::readNWISqw(site, 
#                                        parameterCd)
# Using package example data:
site_data <- L2701_example_data$QW
title <- paste(attr(site_data, "siteInfo")[["station_nm"]], ": Specific Conductance vs Chloride")
Sc_Cl_plot(site_data, title = title)

site <- "263819081585801"
parameterCd <- c("00095","90095","00940","99220")
site_data <- dataRetrieval::readNWISqw(site, 
                                       parameterCd)
sc_cl <- Sc_Cl_table(site_data)
}
