% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_Cl_plot.R
\name{Sc_Cl_plot}
\alias{Sc_Cl_plot}
\title{ScCLPlotTable}
\usage{
Sc_Cl_plot(x, title)
}
\arguments{
\item{x}{data frame}

\item{title}{character}
}
\description{
Function creates the individual chloride versus specfic conductance plots for each site
as well as the chloride versus specific conductance plot for all sites.
}
\examples{

site <- '254457080160301'
site_data <- dataRetrieval::readNWISqw(site, 
                                       parameterCd = c("99220","90095"))
Sc_Cl_plot(site_data, title = "Hi")
}
