% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{site_data_summary}
\alias{site_data_summary}
\title{site_data_summary}
\usage{
site_data_summary(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame with 10 columns
}
\description{
Get summaries of data by site. Requires a column site_no, and will
take the summaries
}
\examples{
aquifer_data <- aquifer_data
summary_info <- site_data_summary(aquifer_data)
}
