% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_utils.R
\name{theme_gwl}
\alias{theme_gwl}
\title{ggplot2 themes and accesories for package}
\usage{
theme_gwl(base_family = "", ...)
}
\arguments{
\item{base_family}{character base font family}

\item{\dots}{additional arguments}
}
\description{
Themes and other useful graphing utilities for the HASP package
}
\examples{



}
