% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trend_test}
\alias{trend_test}
\title{Trend Test}
\usage{
trend_test(
  gw_level_dv,
  gwl_data,
  n_years = 10,
  parameter_cd = NA,
  date_col = NA,
  value_col = NA,
  approved_col = NA,
  stat_cd = NA,
  alpha = 0.95,
  pctComplete = 0.5,
  days_required_per_month = 14,
  POR_trend = TRUE
)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data frame. Often obtained from from \code{readNWISdv}}

\item{gwl_data}{data frame returned from dataRetrieval::readNWISgwl, or 
data frame with mandatory columns lev_dt (representing date), lev_age_cd (representing
approval code), and a column representing the measured value (either lev_va,
sl_lev_va, or value).}

\item{n_years}{integer. This is the number of years to calculate the trend on.
Default is 10. This can be a vector of years.}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd 
can be used to define the value_col.
 If the data doesn't come directly from NWIS services, this 
can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the heading of the date column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{value_col}{name of value column. The default is \code{NA},
which the code will try to get the column name automatically.}

\item{approved_col}{name of column to get provisional/approved status.}

\item{stat_cd}{If data in gw_level_dv comes from NWIS, the stat_cd 
can be used to help define the value_col.}

\item{alpha}{the confidence level to use for statistical significance}

\item{pctComplete}{number percentage complete. This is a fraction that represents
the amount of data that must be included overall in order to calculate a trend.
The default is 0.5, which means if gaps in the data span more than 50% of the
total record, a trend will not be calculated.}

\item{days_required_per_month}{integer. Number of days required per month
to include in the trend test. Default is 14. If discrete data is used, 
it may be required to set this to 0 to get the trend tests to work.}

\item{POR_trend}{a logical indicating whether to include a trend test
for the full period of record. Default is \code{TRUE}.}
}
\value{
a data frame of test results from 5 and 20 year Kendall Seasonal Trend test
}
\description{
Test for five and twenty year trends in groundwater level using the 
Regional Kendall Test.
}
\details{
At least 10 readings per year for the last 5 years are required for the
5-year test, and at least 6 readings for the last 20 years are required
for the 20-year test. The current calendar year is excluded by default.
}
\examples{

# site <- "263819081585801"
# gw_level_data <- dataRetrieval::readNWISgwl(site)

# Using package example data:
gwl_data <- L2701_example_data$Discrete
trend_test(NULL,
           gwl_data,
           parameter_cd = "62610", 
           days_required_per_month = 0)
                        
gw_level_dv <- L2701_example_data$Daily
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610")
           
trend_test(gw_level_dv,
           gwl_data,
           POR_trend = FALSE,
           parameter_cd = "62610")
           
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610",
           n_years = 5)
           
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610",
           n_years = c(5, 10, 20))
           
trend_test(gw_level_dv,
           gwl_data,
           parameter_cd = "62610",
           n_years = 5,
           days_required_per_month = 0)
           
}
