% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_plot}
\alias{weekly_frequency_plot}
\title{Plot weekly frequency analysis}
\usage{
weekly_frequency_plot(
  gw_level_dv,
  parameterCd,
  statCd,
  title = "",
  date = Sys.Date()
)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{parameterCd}{the parameter code used}

\item{statCd}{the statistic code used}

\item{title}{the title to use on the plot}

\item{date}{the reference date to make the plot, defaults to the current date}
}
\value{
a ggplot object with rectangles representing the historical weekly percentiles,
and points representing the historical median and daily values
}
\description{
The weekly frequency analysis is based on daily data
}
\examples{

site <- "263819081585801"
parameterCd <- "62610"
statCd <- "00001"
gw_level_dv <- dataRetrieval::readNWISdv(site, parameterCd, statCd = statCd)
weekly_frequency_plot(gw_level_dv, parameterCd, statCd, title = "Groundwater Level")

}
