% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_plot}
\alias{weekly_frequency_plot}
\title{Plot weekly frequency analysis}
\usage{
weekly_frequency_plot(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status"),
  plot_range = "Past year",
  plot_title = "",
  subtitle = "U.S. Geological Survey",
  y_axis_label = "",
  flip = FALSE,
  percentile_colors = NA
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{gwl_data}{data frame returned from  \code{\link[dataRetrieval]{readNWISgwl}}, or
data frame with a date, value, and approval columns. Using the convention:
lev_dt (representing date), lev_age_cd (representing approval code), and lev_va
or sl_lev_va (representing value) will allow defaults to work.
Use \code{NULL} for no discrete data.}

\item{parameter_cd}{If data in gw_level_dv comes from NWIS, the parameter_cd
can be used to define the value_col. If the data doesn't come directly from
NWIS services, this can be set to \code{NA},and this argument will be ignored.}

\item{date_col}{the name of the date column. The default is \code{NA},
in which case, the code will try to get the column name automatically based on NWIS
naming conventions. If both gw_level_dv and gwl_data data frames
require custom column names, the first value of this input defines the date
column for gw_level_dv, and the second defines gwl_data.}

\item{value_col}{the name of the value column. The default is \code{NA},
in which case, the code will try to get the column name automatically based on NWIS
naming conventions. If both gw_level_dv and gwl_data data frames
require custom column names, the first value of this input defines the value
column for gw_level_dv, and the second defines gwl_data.}

\item{approved_col}{the name of the column to get provisional/approved status.
The default is \code{NA}, in which case, the code will try to get the column name
automatically based on NWIS naming conventions. If both gw_level_dv and
gwl_data data frames require custom column names, the first value of this
input defines the approval column for gw_level_dv, and the second defines gwl_data.}

\item{plot_range}{the time frame to use for the plot. Either "Past year" to use the
last year of data, or "Calendar year" to use the current calendar year, beginning
in January.}

\item{plot_title}{the title to use on the plot}

\item{subtitle}{character. Sub-title for plot, default is "U.S. Geological Survey".}

\item{y_axis_label}{the label used for the y-axis of the plot.}

\item{flip}{logical. If \code{TRUE}, flips the y axis so that the smallest number is on top.
Default is \code{FALSE}.}

\item{percentile_colors}{Optional argument to provide a vector of 5 colors
used to fill the percentile bars in order from lowest percentile bin to the
highest percentile bin. Default behavior (NA) is to use legacy plot colors.}
}
\value{
a ggplot object with rectangles representing the historical weekly percentiles,
and points representing the historical median and daily values
}
\description{
Weekly statistics are calculated using the \code{weekly_frequency_table} function.
Daily, discrete, or both types of data can be used.
}
\examples{

site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site, 
                skipGeometry = TRUE)
site_info <- dataRetrieval::read_waterdata_monitoring_location(monitoring_location_id = site)

gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
                                                   parameter_code = p_code_dv,
                                                   statistic_id = statCd,
                                                   skipGeometry = TRUE)

weekly_frequency_plot(gw_level_dv,
                      gwl_data = NULL)

weekly_frequency_plot(gw_level_dv,
                      gwl_data = gwl_data,
                      parameter_cd = "62610")

weekly_frequency_plot(gw_level_dv,
                      gwl_data = gwl_data,
                      parameter_cd = "62610",
                      flip = TRUE)

}
