% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_plot}
\alias{weekly_frequency_plot}
\title{Plot weekly frequency analysis}
\usage{
weekly_frequency_plot(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status"),
  plot_range = "Past year",
  plot_title = "",
  subtitle = "U.S. Geological Survey",
  y_axis_label = "",
  flip = FALSE,
  percentile_colors = NA
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{gwl_data}{data frame returned from \code{\link[dataRetrieval]{read_waterdata_field_measurements}}, or
data frame with a date, value, and approval columns. 
Use \code{NULL} for no discrete data.}

\item{parameter_cd}{Can be used to filter data if the data frame has a 
"parameter_code" column. The default is \code{NA},
which will not do any filtering.}

\item{date_col}{the name of the time columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("time", "time")}.}

\item{value_col}{the name of the value columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("value", "value")}.}

\item{approved_col}{the name of the column to get provisional/approved status. 
The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("approval_status", "approval_status")}.
It is expected that these columns will have only "Approved" or "Provisional".}

\item{plot_range}{the time frame to use for the plot. Either "Past year" to use the
last year of data, or "Calendar year" to use the current calendar year, beginning
in January.}

\item{plot_title}{the title to use on the plot}

\item{subtitle}{character. Sub-title for plot, default is "U.S. Geological Survey".}

\item{y_axis_label}{the label used for the y-axis of the plot.}

\item{flip}{logical. If \code{TRUE}, flips labels so that the lower numbers
are in the higher percentages. Default is \code{TRUE}.}

\item{percentile_colors}{Optional argument to provide a vector of 5 colors
used to fill the percentile bars in order from lowest percentile bin to the
highest percentile bin. Default behavior (NA) is to use legacy plot colors.}
}
\value{
a ggplot object with rectangles representing the historical weekly percentiles,
and points representing the historical median and daily values
}
\description{
Weekly statistics are calculated using the \code{weekly_frequency_table} function.
Daily, discrete, or both types of data can be used.
}
\examples{

site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"

# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site, 
#                                  skipGeometry = TRUE)
gwl_data <- L2701_example_data$Discrete

# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd,
#                                                    skipGeometry = TRUE)
#
                                                    
gw_level_dv <- L2701_example_data$Daily

weekly_frequency_plot(gw_level_dv,
                      gwl_data = NULL)
                      
weekly_frequency_plot(gw_level_dv,
                      gwl_data = gwl_data,
                      parameter_cd = "62610",
                      plot_range = c("2023-10-01", "2024-09-30"))

weekly_frequency_plot(gw_level_dv,
                      gwl_data = gwl_data,
                      parameter_cd = "62610")

weekly_frequency_plot(gw_level_dv,
                      gwl_data = gwl_data,
                      parameter_cd = "62610",
                      flip = TRUE)

}
