% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_table}
\alias{weekly_frequency_table}
\title{Create a table of weekly frequency analysis}
\usage{
weekly_frequency_table(gw_level_dv, p_code_dv, statCd)
}
\arguments{
\item{gw_level_dv}{daily groundwater level data
from readNWISdv}

\item{p_code_dv}{the parameter code used}

\item{statCd}{the statistic code used}
}
\value{
a data frame of weekly frequency analysis
}
\description{
The weekly frequency analysis is based on daily values
}
\examples{

# site <- "263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::readNWISdv(site, p_code_dv, statCd = statCd)
gw_level_dv <- L2701_example_data$Daily
weekly_frequency <- weekly_frequency_table(gw_level_dv, p_code_dv, statCd)

}
