% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{weekly_frequency_table}
\alias{weekly_frequency_table}
\title{Create a table of weekly frequency analysis}
\usage{
weekly_frequency_table(
  gw_level_dv,
  gwl_data,
  parameter_cd = NA,
  date_col = c("time", "time"),
  value_col = c("value", "value"),
  approved_col = c("approval_status", "approval_status"),
  flip = FALSE
)
}
\arguments{
\item{gw_level_dv}{data frame, daily groundwater level data. Often obtained
from \code{\link[dataRetrieval]{read_waterdata_daily}}. Use \code{NULL} for no daily data.}

\item{gwl_data}{data frame returned from \code{\link[dataRetrieval]{read_waterdata_field_measurements}}, or
data frame with a date, value, and approval columns. 
Use \code{NULL} for no discrete data.}

\item{parameter_cd}{Can be used to filter data if the data frame has a 
"parameter_code" column. The default is \code{NA},
which will not do any filtering. If the gwl_data and gw_level_dv need different
parameter code filtering, use a vector of 2 parameter codes. The first one will filter the
gw_level_dv data frame, the second will filter the gwl_data data frame.}

\item{date_col}{the name of the time columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("time", "time")}.}

\item{value_col}{the name of the value columns. The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("value", "value")}.}

\item{approved_col}{the name of the column to get provisional/approved status. 
The first value is associated
with the gw_level_dv input, and the second value is associated with the gwl_data
input. The default is \code{c("approval_status", "approval_status")}.
It is expected that these columns will have only "Approved" or "Provisional".}

\item{flip}{logical. If \code{TRUE}, flips labels so that the lower numbers
are in the higher percentages. Default is \code{TRUE}.}
}
\value{
a data frame of weekly frequency analysis
}
\description{
The weekly frequency analysis is based on daily, discrete, or both types of
data. The median of each year/week combo is calculated, then overall weekly
statistics are calculated off of that median.
}
\examples{

site <- "USGS-263819081585801"
p_code_dv <- "62610"
statCd <- "00001"
# gw_level_dv <- dataRetrieval::read_waterdata_daily(monitoring_location_id = site,
#                                                    parameter_code = p_code_dv,
#                                                    statistic_id = statCd)
                                                    
gw_level_dv <- L2701_example_data$Daily

weekly_frequency <- weekly_frequency_table(gw_level_dv, NULL)
head(weekly_frequency)

# gwl_data <- dataRetrieval::read_waterdata_field_measurements(monitoring_location_id = site)
gwl_data <- L2701_example_data$Discrete

weekly_frequency <- weekly_frequency_table(gw_level_dv,
                                           gwl_data,
                                           parameter_cd = "62610")
head(weekly_frequency)
weekly_flip <- weekly_frequency_table(gw_level_dv,
                                      gwl_data,
                                      parameter_cd = "62610",
                                      flip = TRUE)
head(weekly_flip)
}
